pyproject.toml
[project]
name = "oraculum"
version = "0.1.0"
description = "Oráculo contemporâneo — instalação de arte interativa"
requires-python = ">=3.10,<3.14"
dependencies = [
    "opencv-python>=4.8",
    "sounddevice>=0.4",
    "numpy>=1.24",
    "pyobjc-framework-AVFoundation>=10.0; sys_platform == 'darwin'",
]

[tool.uv]
python-preference = "managed"
